style(client-handler): update termination condition in 'run' method
In the 'run' method of the 'ClientHandler' class, the termination condition of the while loop has been updated. The loop will now terminate when there are no more available bytes to read from the input stream. This change resolves the issue associated with EOFException, ensuring the loop doesn't continue unnecessarily.